No new functionality, per se, just a lot of cleanup.
The modules (.pm files) are now really modules, rather than just library
scripts.  They also "use script" and "use warnings".

The one small piece of functionality I added was to redirect CGI error
messages from the central Apache server to a CIVS error log named
cgi-log.

Of course, adding both of the above meant that I started seeing lots
of warnings in the log, so I fixed all the bugs that were causing
those warnings.  And there was lots of stuff to fix because of turning
on strictness.

Now that the modules properly export their interface, the way is clear
for adding strictness and warnings to the CGI scripts, since they no longer
have to rely on global variables.   Actually, that's the reason I did all
this to begin with.  But that's a project for another weekend.