Implemented the lastest changes that Andrew and I designed.  The important changes are that user keys are now deterministic and control keys are now random.
User keys are constructed as h(email,auth_key,private_host_id).  This
ensures that a voter can't vote (with the same email address)
more than once, even if the voter requests to be authorized a second time.
It means we get to get rid of %authorized_voters, since the voter key
is deterministic.

Control keys are now nonces.  We store only the hash of the key,
and compare to verify it.  If no hash is stored, we grandfather in the
election and compute the control key check as we used to.  Go backwards
compatibility.

We can't map from voter keys to ballots now, since supervisors can
figure out voter keys if they hack the server.  So it's now a map
from a voter key and nonce to the ballot.  Later, we can use that
nonce as part of ballot confirmation, if we decide to implement that.
Because of this, we have to keep track of used voter keys separately.
So we have a new hash %used_voter_keys for this.

Also, we were previously storing a list of authorized voter keys,
so we could check whether a user-supplied key was indeed authorized.
But that means someone who hacks the server could get that list
and vote on behalf of those voters.  Bad.  So now we store
the hash of each key, and compare to verify.

Some smaller things:
* created &civs_hash() to replace the frequent calls to substr(md5(...))
* tweaked text on control page
* added a status script.  This is *not* installed by default; it's
  just for testing/information purposes.

I tested all this with the '04 POTUS public election.  Seems good.