Actions now have a notion of equivalence. (cf. Action.equivalentTo())
A canonical representative for a given equivalence class of
Actions can be obtained by calling Action.intern().

To facilitate Action object reuse, the constructors for
Hyperlink and Form now intern() the actions passed to their
respective constructors.

This allows the "Create Event" link on a Calendar to be
constant in a given session.  This means that if a user uses
their browser's "Back" button to go back to a Calendar view,
and clicks on the "Create new Event" link, the link will
work.