Added principals to the servlet framework.
The key design idea is to allow both container-based authentication and application-specific authentication to be incorporated into the DLM's notion of principals.

Towards this end, we have a session principals (representing the http sessions), user principals (representing the container's notion of users) and role principals (representing the container's notion of roles). When a user is authenticated using container-based authentication, the session principal is automatically regarded as acting for the user principal. A user principal is regarded as acting for a role principal if the container asserts that the user has that role. Some ugly hacks were needed to hide away the container's notion of authentication and role checking behind a DLM principal interface.

Application specific authentication will require the applications to be able to get their hands on appropriate principal objects, and add actsfor relations as needed. For example, if a client presents a voter key, that should be sufficient to create a principal for the voter key, and allow the session principal to actfor the voter-key principal. We need to carefully figure out how we restrict web servlet code from adding arbitrary actsfor relationships.

The calendar app now requires basic authentication. We can explore using more complicated authentication schemes later.