streamline styles
Font change 1: Everyone agrees Courier New is far too thin. If you want a Courier, Prime is the ideal. On Windows, a request for Courier is silently replaced with Courier New anyway, so that's what you'll get. You don't need to specify it. This allows better Couriers to be used on systems (like Mac) that probably have one.

Font change 2: Everyone agrees Arial is a shoddy copy of Helvetica. On Windows (where Helv is not typically installed), a request for Helvetica is silently replaced with Arial anyway, so that's what you'll get (even if you do have Helv). You don't need to specify it. This allows platforms that have Helvetica (like Mac) to use it rather than the knock-off.

Width change: The main content was displayed in far too narrow a space, wasting lots of room on margins. Allow this to be wider, but use columns to enforce a readable line length.